% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctk_packages.R
\name{ctk_packages}
\alias{ctk_packages}
\title{Package List}
\usage{
ctk_packages()
}
\value{
A data frame of package metadata.
}
\description{
Return the embedded package list copied from \url{https://christophertkenny.com}.
}
\examples{
ctk_packages()
}
